Struct forked_pallet_democracy::Module

pub struct Module<T: Trait>(_);

Implementations

impl<T: Trait + 'static> Module<T>

pub fn public_prop_count() -> PropIndex

The number of (public) proposals that have been made so far.

pub fn public_props() -> Vec<(PropIndex, T::Hash, T::AccountId)>

The public proposals. Unsorted. The second item is the proposal’s hash.

pub fn deposit_of<K: EncodeLike<PropIndex>>(
key: K
) -> Option<(Vec<T::AccountId>, BalanceOf<T>)>

Those who have locked a deposit.

TWOX-NOTE: Safe, as increasing integer keys are safe.

pub fn referendum_count() -> ReferendumIndex

The next free referendum index, aka the number of referenda started so far.

pub fn lowest_unbaked() -> ReferendumIndex

The lowest referendum index representing an unbaked referendum. Equal to ReferendumCount if there isn’t an unbaked referendum.

pub fn referendum_info<K: EncodeLike<ReferendumIndex>>(
key: K
) -> Option<ReferendumInfo<T::BlockNumber, T::Hash, BalanceOf<T>>>

Information concerning any given referendum.

TWOX-NOTE: SAFE as indexes are not under an attacker’s control.

pub fn locks<K: EncodeLike<T::AccountId>>(key: K) -> Option<T::BlockNumber>

Accounts for which there are locks in action which may be removed at some point in the future. The value is the block number at which the lock expires and may be removed.

TWOX-NOTE: OK ― AccountId is a secure hash.

pub fn blacklist<K: EncodeLike<T::Hash>>(
key: K
) -> Option<(T::BlockNumber, Vec<T::AccountId>)>

A record of who vetoed what. Maps proposal hash to a possible existent block number (until when it may not be resubmitted) and who vetoed it.

impl<T: Trait> Module<T>

Can also be called using Call.

pub fn propose(
origin: T::Origin,
proposal_hash: T::Hash,
value: BalanceOf<T>
) -> DispatchResult

Propose a sensitive action to be taken.

The dispatch origin of this call must be Signed and the sender must have funds to cover the deposit.

proposal_hash: The hash of the proposal preimage.

value: The amount of deposit (must be at least MinimumDeposit).

Emits Proposed.

Complexity: O(1)

Db reads: PublicPropCount, PublicProps

Db writes: PublicPropCount, PublicProps, DepositOf

NOTE: Calling this function will bypass origin filters.

pub fn second(
origin: T::Origin,
proposal: PropIndex,
seconds_upper_bound: u32
) -> DispatchResult

Signals agreement with a particular proposal.

The dispatch origin of this call must be Signed and the sender must have funds to cover the deposit, equal to the original deposit.

proposal: The index of the proposal to second.

seconds_upper_bound: an upper bound on the current number of seconds on this proposal. Extrinsic is weighted according to this value with no refund.

Complexity: O(S) where S is the number of seconds a proposal already has.

Db reads: DepositOf

Db writes: DepositOf

NOTE: Calling this function will bypass origin filters.

pub fn emergency_cancel(
origin: T::Origin,
ref_index: ReferendumIndex
) -> DispatchResult

Schedule an emergency cancellation of a referendum. Cannot happen twice to the same referendum.

The dispatch origin of this call must be CancellationOrigin.

-ref_index: The index of the referendum to cancel.

Complexity: O(1).

Db reads: ReferendumInfoOf, Cancellations

Db writes: ReferendumInfoOf, Cancellations

NOTE: Calling this function will bypass origin filters.

pub fn external_propose_majority(
origin: T::Origin,
proposal_hash: T::Hash
) -> DispatchResult

Schedule a majority-carries referendum to be tabled next once it is legal to schedule an external referendum.

The dispatch of this call must be ExternalMajorityOrigin.

proposal_hash: The preimage hash of the proposal.

Unlike external_propose, blacklisting has no effect on this and it may replace a pre-scheduled external_propose call.

Complexity: O(1)

Db write: NextExternal

NOTE: Calling this function will bypass origin filters.

pub fn fast_track(
origin: T::Origin,
proposal_hash: T::Hash,
voting_period: T::BlockNumber,
delay: T::BlockNumber
) -> DispatchResult

Schedule the currently externally-proposed majority-carries referendum to be tabled immediately. If there is no externally-proposed referendum currently, or if there is one but it is not a majority-carries referendum then it fails.

The dispatch of this call must be FastTrackOrigin.

proposal_hash: The hash of the current external proposal.

voting_period: The period that is allowed for voting on this proposal. Increased to FastTrackVotingPeriod if too low.

delay: The number of block after voting has ended in approval and this should be enacted. This doesn’t have a minimum amount.

Emits Started.

Complexity: O(1)

Db reads: NextExternal, ReferendumCount

Db writes: NextExternal, ReferendumCount, ReferendumInfoOf

Base Weight: 30.1 µs

NOTE: Calling this function will bypass origin filters.

pub fn cancel_referendum(
origin: T::Origin,
ref_index: ReferendumIndex
) -> DispatchResult

Remove a referendum.

The dispatch origin of this call must be Root.

ref_index: The index of the referendum to cancel.

Complexity: O(1).

Db writes: ReferendumInfoOf

NOTE: Calling this function will bypass origin filters.

pub fn cancel_queued(
origin: T::Origin,
which: ReferendumIndex
) -> DispatchResult

Cancel a proposal queued for enactment.

The dispatch origin of this call must be Root.

which: The index of the referendum to cancel.

O(D) where D is the items in the dispatch queue. Weighted as D = 10.

Db reads: scheduler lookup, scheduler agenda`

Db writes: scheduler lookup, scheduler agenda`

NOTE: Calling this function will bypass origin filters.

pub fn delegate(
origin: T::Origin,
to: T::AccountId,
conviction: Conviction,
balance: BalanceOf<T>
) -> DispatchResultWithPostInfo

Delegate the voting power (with some given conviction) of the sending account.

The balance delegated is locked for as long as it’s delegated, and thereafter for the time appropriate for the conviction’s lock period.

The dispatch origin of this call must be Signed, and the signing account must either:

be delegating already; or

have no voting activity (if there is, then it will need to be removed/consolidated through reap_vote or unvote).

to: The account whose voting the target account’s voting power will follow.

conviction: The conviction that will be attached to the delegated votes. When the account is undelegated, the funds will be locked for the corresponding period.

balance: The amount of the account’s balance to be used in delegating. This must not be more than the account’s current balance.

Emits Delegated.

Complexity: O(R) where R is the number of referendums the voter delegating to has voted on. Weight is charged as if maximum votes.

Db reads: 3*VotingOf, origin account locks

Db writes: 3*VotingOf, origin account locks

Db reads per votes: ReferendumInfoOf

Db writes per votes: ReferendumInfoOf

NOTE: Calling this function will bypass origin filters.

pub fn clear_public_proposals(origin: T::Origin) -> DispatchResult

Clears all public proposals.

The dispatch origin of this call must be Root.

O(1).

Db writes: PublicProps

NOTE: Calling this function will bypass origin filters.

pub fn note_preimage(
origin: T::Origin,
encoded_proposal: Vec<u8>
) -> DispatchResult

Register the preimage for an upcoming proposal. This doesn’t require the proposal to be in the dispatch queue but does require a deposit, returned once enacted.

The dispatch origin of this call must be Signed.

encoded_proposal: The preimage of a proposal.

Emits PreimageNoted.

Complexity: O(E) with E size of encoded_proposal (protected by a required deposit).

Db reads: Preimages

Db writes: Preimages

NOTE: Calling this function will bypass origin filters.

pub fn note_preimage_operational(
origin: T::Origin,
encoded_proposal: Vec<u8>
) -> DispatchResult

Same as note_preimage but origin is OperationalPreimageOrigin.

NOTE: Calling this function will bypass origin filters.

pub fn note_imminent_preimage(
origin: T::Origin,
encoded_proposal: Vec<u8>
) -> DispatchResultWithPostInfo

Register the preimage for an upcoming proposal. This requires the proposal to be in the dispatch queue. No deposit is needed. When this call is successful, i.e. the preimage has not been uploaded before and matches some imminent proposal, no fee is paid.

The dispatch origin of this call must be Signed.

encoded_proposal: The preimage of a proposal.

Emits PreimageNoted.

Complexity: O(E) with E size of encoded_proposal (protected by a required deposit).

Db reads: Preimages

Db writes: Preimages

NOTE: Calling this function will bypass origin filters.

pub fn reap_preimage(
origin: T::Origin,
proposal_hash: T::Hash,
proposal_len_upper_bound: u32
) -> DispatchResult

Remove an expired proposal preimage and collect the deposit.

The dispatch origin of this call must be Signed.

proposal_hash: The preimage hash of a proposal.

proposal_length_upper_bound: an upper bound on length of the proposal. Extrinsic is weighted according to this value with no refund.

This will only work after VotingPeriod blocks from the time that the preimage was noted, if it’s the same account doing it. If it’s a different account, then it’ll only work an additional EnactmentPeriod later.

Emits PreimageReaped.

Complexity: O(D) where D is length of proposal.

Db reads: Preimages, provider account data

Db writes: Preimages provider account data

NOTE: Calling this function will bypass origin filters.

pub fn enact_proposal(
origin: T::Origin,
proposal_hash: T::Hash,
index: ReferendumIndex
) -> DispatchResult

Enact a proposal from a referendum. For now we just make the weight be the maximum.

NOTE: Calling this function will bypass origin filters.

pub fn cancel_proposal(
origin: T::Origin,
prop_index: PropIndex
) -> DispatchResult

Remove a proposal.

The dispatch origin of this call must be CancelProposalOrigin.

prop_index: The index of the proposal to cancel.

Weight: O(p) where p = PublicProps::<T>::decode_len()

NOTE: Calling this function will bypass origin filters.

impl<T: Trait> Module<T>

pub fn backing_for(proposal: PropIndex) -> Option<BalanceOf<T>>

Get the amount locked in support of proposal; None if proposal isn’t a valid proposal index.

pub fn maturing_referenda_at(
n: T::BlockNumber
) -> Vec<(ReferendumIndex, ReferendumStatus<T::BlockNumber, T::Hash, BalanceOf<T>>)>

Get all referenda ready for tally at block n.

pub fn internal_start_referendum(
proposal_hash: T::Hash,
threshold: VoteThreshold,
delay: T::BlockNumber
) -> ReferendumIndex

Start a referendum.

pub fn internal_cancel_referendum(ref_index: ReferendumIndex)

Remove a referendum.

pub fn referendum_status(
ref_index: ReferendumIndex
) -> Result<ReferendumStatus<T::BlockNumber, T::Hash, BalanceOf<T>>, DispatchError>